move to snapping/common.tsx
